requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==2.15.8
astunparse==1.6.3
attrs==24.2.0
black==23.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
elastic-transport==8.15.1
elasticsearch==8.16.0
fake-useragent==1.5.1
Faker==18.13.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
glassflow==1.0.0
google-analytics-data==0.18.15
google-api-core==2.23.0
google-auth==2.36.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
keras==3.6.0
lazy-object-proxy==1.10.0
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.13.4
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.55.0
openai-whisper==20240930
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.3.6
plotly==5.15.0
pluggy==1.5.0
pocketsphinx==5.0.3
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pylint==2.17.5
Pympler==1.1
PySocks==1.7.1
pyspark==3.5.3
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==308
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
selenium==4.26.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SpeechRecognition==3.11.0
streamlit==1.24.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==1.26.15
validators==0.34.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.0
zipp==3.21.0
